requirements.txt
# Core GUI Framework
PyQt5==5.15.11

# Audio Processing
sounddevice==0.5.5
soundfile==0.13.1
pydub==0.25.1
pygame==2.6.1

# Speech Recognition Engines
SpeechRecognition==3.14.5
openai-whisper==20250625

# Text-to-Speech
gTTS==2.5.4

# Scientific Computing
numpy==2.3.5
